Fix concurrency issue on user create/update
When creating or updating a mssql user in servers with compatibility
level below 130, the `string_split` function needs to be created.

At the moment, this is causing issues when trying to create/update
multiple users, as each of those requests tries to create the same
function.

This change introduces a lock on application resource in exclusive
mode so that only one of the calls creates the function.

Fixes #31